Rename and move modules to make docs easier to follow
