+++ firmware/src/adc_capture/timing.rs
